Module asm::asm::core

Core Assembly Variant

This variant of the assembly language is intended to be used with the core variant of the virtual machine. It is extremely portable, but minimal.

What kinds of instructions are supported by this variant?

This variant attempts to support as many instructions as possible that can be implemented WITHOUT the standard virtual machine variant. This includes instructions for operations like Copy (a memcpy clone), static stack allocation, Swap (which uses a TMP register without the more optimized standard Swap instruction), and DivMod, which performs a division and modulo operation in a single instruction.

What kinds of instructions are not supported by this variant?

Mainly, this variant is lacking in I/O instructions and memory allocation instructions. This is because of the bare bones core virtual machine specification which only includes 2 I/O instructions (GetChar and PutChar), and does not include any memory allocation instructions.

Standard instructions, like PutInt, can be implemented as user defined functions in the core assembly language simply using PutChar to display the integer in decimal.

Structs

CoreProgram

A program composed of core instructions, which can be assembled into the core virtual machine instructions.

Enums

CoreOp

A core instruction of the assembly language. These are instructions guaranteed to be implemented for every target possible.