Enum asm::asm::std::StandardOp

pub enum StandardOp {

    Comment(String),
    Set(Location, isize),
    SetLabel(Location, String),
    GetAddress {
        addr: Location,
        dst: Location,
    },
    Call(Location),
    CallLabel(String),
    Return,
    Fn(String),
    While(Location),
    If(Location),
    Else,
    End,
    Move {
        src: Location,
        dst: Location,
    },
    Swap(Location, Location),
    Next(Location, Option<isize>),
    Prev(Location, Option<isize>),
    Inc(Location),
    Dec(Location),
    Add {
        src: Location,
        dst: Location,
    },
    Sub {
        src: Location,
        dst: Location,
    },
    Mul {
        src: Location,
        dst: Location,
    },
    Div {
        src: Location,
        dst: Location,
    },
    Rem {
        src: Location,
        dst: Location,
    },
    DivRem {
        src: Location,
        dst: Location,
    },
    Neg(Location),
    Not(Location),
    And {
        src: Location,
        dst: Location,
    },
    Or {
        src: Location,
        dst: Location,
    },
    Push(Location),
    Pop(Option<Location>),
    Compare {
        dst: Location,
        a: Location,
        b: Location,
    },
    IsGreater {
        dst: Location,
        src: Location,
    },
    IsGreaterEqual {
        dst: Location,
        src: Location,
    },
    IsLess {
        dst: Location,
        src: Location,
    },
    IsLessEqual {
        dst: Location,
        src: Location,
    },
    IsEqual {
        dst: Location,
        src: Location,
    },
    IsNotEqual {
        dst: Location,
        src: Location,
    },
    GetChar(Location),
    PutChar(Location),

}

A core instruction of the assembly language. These are instructions guaranteed to be implemented for every target possible.

Variants

Comment(String)

Tuple Fields of Comment

0: String

Set(Location, isize)

Set the value of a register, or any location in memory, to a given constant.

Tuple Fields of Set

0: Location1: isize

SetLabel(Location, String)

Set the value of a register, or any location in memory, to the value of a label’s ID.

Tuple Fields of SetLabel

0: Location1: String

GetAddress

Get the address of a location, and store it in a destination

Fields of GetAddress

addr: Locationdst: Location

Call(Location)

Get a value in memory and call it as a label ID.

Tuple Fields of Call

0: Location

CallLabel(String)

Call a function with a given label.

Tuple Fields of CallLabel

0: String

Return

Return from the current function.

Fn(String)

Declare a new label.

Tuple Fields of Fn

0: String

While(Location)

Begin a “while the value is not zero” loop over a given register or location in memory.

Tuple Fields of While

0: Location

If(Location)

Begin an “if the value is not zero” statement over a given register or location in memory.

Tuple Fields of If

0: Location

Else

Add an “else” clause to an “if the value is not zero” statement.

End

Terminate a function declaration, a while loop, an if statement, or an else clause.

Move

Copy a value from a source location to a destination location.

Fields of Move

src: Locationdst: Location

Swap(Location, Location)

Swap the values of two locations.

Tuple Fields of Swap

0: Location1: Location

Next(Location, Option<isize>)

Make this pointer point to the next cell.

Tuple Fields of Next

0: Location1: Option<isize>

Prev(Location, Option<isize>)

Make this pointer point to the previous cell.

Tuple Fields of Prev

0: Location1: Option<isize>

Inc(Location)

Increment the integer value of a location.

Tuple Fields of Inc

0: Location

Dec(Location)

Decrement the integer value of a location.

Tuple Fields of Dec

0: Location

Add

Add an integer value from a source location to a destination location.

Fields of Add

src: Locationdst: Location

Sub

Subtract a source integer value from a destination location.

Fields of Sub

src: Locationdst: Location

Mul

Multiply a destination location by a source value.

Fields of Mul

src: Locationdst: Location

Div

Divide a destination location by a source value.

Fields of Div

src: Locationdst: Location

Rem

Store the remainder of the destination modulus the source in the destination.

Fields of Rem

src: Locationdst: Location

DivRem

Divide a destination location by a source value. Store the quotient in the destination, and the remainder in the source.

Fields of DivRem

src: Locationdst: Location

Neg(Location)

Negate an integer.

Tuple Fields of Neg

0: Location

Not(Location)

Replace a value in memory with its boolean complement.

Tuple Fields of Not

0: Location

And

Logical “and” a destination with a source value.

Fields of And

src: Locationdst: Location

Or

Logical “or” a destination with a source value.

Fields of Or

src: Locationdst: Location

Push(Location)

Push an integer at a memory location on the stack.

Tuple Fields of Push

0: Location

Pop(Option<Location>)

Pop an integer from the stack and store it in a memory location.

Tuple Fields of Pop

0: Option<Location>

Compare

Store the comparison of “a” and “b” in a destination register. If “a” is less than “b”, store -1. If “a” is greater than “b”, store 1. If “a” is equal to “b”, store 0.

Fields of Compare

dst: Locationa: Locationb: Location

IsGreater

Perform dst = dst > src.

Fields of IsGreater

dst: Locationsrc: Location

IsGreaterEqual

Perform dst = dst >= src.

Fields of IsGreaterEqual

dst: Locationsrc: Location

IsLess

Perform dst = dst < src.

Fields of IsLess

dst: Locationsrc: Location

IsLessEqual

Perform dst = dst <= src.

Fields of IsLessEqual

dst: Locationsrc: Location

IsEqual

Perform dst = dst == src.

Fields of IsEqual

dst: Locationsrc: Location

IsNotEqual

Perform dst = dst != src.

Fields of IsNotEqual

dst: Locationsrc: Location

GetChar(Location)

Get a character from the input stream and store it in a destination register.

Tuple Fields of GetChar

0: Location

PutChar(Location)

Put a character from a source register to the output stream.

Tuple Fields of PutChar

0: Location